+++ assets/javascripts/kitten/karl/editor/components/button.js
@@ -1,9 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import {
-  Button,
-  COLORS,
-} from 'kitten'
+import { Button, COLORS } from 'kitten'
 import { ICON } from '../../../components/action/button/standalone-styles'
 import { modifierStyles } from '../../../components/action/button/helpers/modifier-styles'