+++ assets/javascripts/kitten/karl/pages/kissbankers-modal/components/group.js
@@ -1,11 +1,7 @@
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import {
-  pxToRem,
-  TYPOGRAPHY,
-  ScreenConfig,
-} from 'kitten'
+import { pxToRem, TYPOGRAPHY, ScreenConfig } from 'kitten'
 import { containerStyles, titleStyles } from './styles'
 import {
   AdaptableGrid as Grid,