+++ src/components/structure/carousels/carousel/components/carousel-inner.js
@@ -165,10 +165,7 @@ var CarouselInner = function CarouselInner(_ref) {
     var isActivePage = currentPageIndex === index;
     var hasPageBeenViewed = viewedPages.has(index);
     return /*#__PURE__*/_react.default.createElement("div", {
-      key: index,
-      role: "button",
-      "aria-label": pageClickText(index + 1),
-      onClick: handlePageClick(index),
+      key: "inner_" + index,
       className: (0, _classnames.default)('k-Carousel__inner__pageContainer', pagesClassName, {
         'k-Carousel__inner__pageContainer--isActivePage': isActivePage,
         'k-Carousel__inner__pageContainer--hasBeenViewed': hasPageBeenViewed
@@ -182,6 +179,11 @@ var CarouselInner = function CarouselInner(_ref) {
       goToCurrentPage: function goToCurrentPage() {
         return goToPage(index);
       }
+    }), !isActivePage && /*#__PURE__*/_react.default.createElement("button", {
+      type: "button",
+      onClick: handlePageClick(index),
+      className: "k-u-reset-button k-Carousel__inner__button",
+      "aria-label": pageClickText(index + 1)
     }));
   }));
 };